+++ msg_hash.h
@@ -1,7 +1,7 @@
 /*  RetroArch - A frontend for libretro.
  *  Copyright (C) 2010-2014 - Hans-Kristian Arntzen
  *  Copyright (C) 2011-2017 - Daniel De Matteis
- *  Copyright (C) 2016-2017 - Brad Parker
+ *  Copyright (C) 2016-2019 - Brad Parker
  *
  *  RetroArch is free software: you can redistribute it and/or modify it under the terms
  *  of the GNU General Public License as published by the Free Software Found-
@@ -2230,6 +2230,7 @@ enum msg_hash_enums
 #endif
 
    MENU_ENUM_LABEL_VALUE_HOLD_START,
+   MENU_LABEL(PLAYLIST_USE_OLD_FORMAT),
 
    MSG_LAST
 };