+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 1.7.6 (future)
 - ANDROID: Fix Xperia Play input binding
 - CHEEVOS: Reset when hardcore mode is toggled
+- COMMON: Add new JSON playlist format
 - CORE UPDATER: Allow sideloading cores from the menu
 - CPU FILTERS: Add Normal2x filter.
 - DATE: Add Date / Time style options.